+++ scanner/src/sockets_iter.rs
@@ -19,9 +19,8 @@ impl<'s> Iterator for SocketIterator<'s> {
     type Item = SocketAddr;
 
     fn next(&mut self) -> Option<Self::Item> {
-        match self.product_it.next() {
-            None => None,
-            Some((port, ip)) => Some(SocketAddr::new(*ip, *port)),
-        }
+        self.product_it
+            .next()
+            .map(|(port, ip)| SocketAddr::new(*ip, *port))
     }
 }